refactor(checker): split class duplicate-member checks out of interface_checks
Extract the ~820-line class duplicate-member checking section
(report_duplicate_class_member_ts2300, get_class_*_info helpers, and
check_duplicate_class_members) from interface_checks.rs into a new sibling
class_member_checks.rs. The section validates class member declarations
(TS2300/TS2393/TS2717) and is a distinct concern from the interface
declaration/variance/duplicate-interface-member checks the file is named for.

Behavior-preserving move: the only logic changes are widening two helpers
(should_check_late_bound_class_property_name, is_late_bound_member_name) from
private fn to pub(crate) fn so the moved code can still call them across the
module boundary. check_duplicate_class_members stays pub(crate) and its
external callers (class.rs) are unaffected.

interface_checks.rs: 2250 -> 1438 LOC (now under the 2000-line checker
boundary). Removes its entry from the arch-guard LINE_LIMIT_CHECKS allowlist
and drops the stale assignability_diagnostics.rs entry (already 1998 LOC),
so test_excluded_files_actually_exceed_limit passes for the checker set.

AgentName: M1-Opus
